phase 8: gather + build + destroy slice
Foundation: positions across game_core, game_persistence, and the
snapshot wire switch from floats to scaled integers (1000 sub-units
per world unit); frontend divides by 1000 at the channel boundary so
Three.js stays float-native.

Catalogue (closed compile-time enums): Item (:wood), Worldgen places
five trees clustered at chunk-centre, Structure.Catalogue (:wall costs
5 wood, 100 HP, 25 damage per click).

Verbs on GameCore.Chunk — harvest/3, build/4, damage/3 — all bounded
by one @interact_range_sq (1.0 world unit). Build is atomic via a
single Repo.transaction (INSERT structure + UPDATE inventory). Damage
deletes the row + ECS entity when HP hits zero.

Wire protocol: snapshot extends to parallel keys
%{players, resource_nodes, structures}, and inventory deltas publish
to a per-owner PubSub topic "self:<username>" that only the owner
channel subscribes to.

Frontend: plain-DOM #inv-hud renders inventory counts on `self` push;
click handler raycasts and target-infers harvest / damage / build,
with explicit verb-pushers on `__game` for the e2e test.

Deferred: tree respawn timer + depletion-state persistence.

+++ apps/game_persistence/lib/game_persistence/schemas/resource_node.ex
@@ -8,8 +8,8 @@ defmodule GamePersistence.Schemas.ResourceNode do
     field(:chunk_x, :integer)
     field(:chunk_y, :integer)
     field(:type, :string)
-    field(:x, :float)
-    field(:y, :float)
+    field(:x, :integer)
+    field(:y, :integer)
     field(:depleted_until, :utc_datetime_usec)
 
     timestamps(type: :utc_datetime_usec)
